int的取值范围为（-2147483648~2147483647），占用4个字节（-2的31次方到2的31次方-1）
long的取值范围为（-9223372036854774808~9223372036854774807），占用8个字节（-2的63次方到2的63次方-1）

+++ src/test/kotlin/com/ak47/cms/cms/TimestampTest.kt
@@ -10,7 +10,7 @@ class TimestampTest {
 
     @Test
     fun test(){
-        val time= 1511850397L * 1000
+        val time= 1511850397.toLong() * 1000
         val format = SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
         val d = format.format(time)
         println(d)
